# gateway/api/gateway/v1/dataVisualization_router.pb.go
// Code generated by https://github.com/zhufuyi/sponge, DO NOT EDIT.

package v1

import (
	context "context"
	gin "github.com/gin-gonic/gin"
	errcode "github.com/zhufuyi/sponge/pkg/errcode"
	middleware "github.com/zhufuyi/sponge/pkg/gin/middleware"
	zap "go.uber.org/zap"
	strings "strings"
)

// import packages: strings. context. errcode. middleware. zap. gin.

type DataVisualizationLogicer interface {
	DeviceDataGet(ctx context.Context, req *DeviceDataGetReq) (*DeviceDataGetRes, error)
	DeviceDataPush(ctx context.Context, req *DeviceDataPushReq) (*DeviceDataPushRes, error)
	DeviceDataCurve(ctx context.Context, req *DeviceDataCurveReq) (*DeviceDataCurveRes, error)
}

type DataVisualizationOption func(*dataVisualizationOptions)

type dataVisualizationOptions struct {
	isFromRPC  bool
	responser  errcode.Responser
	zapLog     *zap.Logger
	httpErrors []*errcode.Error
	rpcStatus  []*errcode.RPCStatus
	wrapCtxFn  func(c *gin.Context) context.Context
}

func (o *dataVisualizationOptions) apply(opts ...DataVisualizationOption) {
	for _, opt := range opts {
		opt(o)
	}
}

func WithDataVisualizationHTTPResponse() DataVisualizationOption {
	return func(o *dataVisualizationOptions) {
		o.isFromRPC = false
	}
}

func WithDataVisualizationRPCResponse() DataVisualizationOption {
	return func(o *dataVisualizationOptions) {
		o.isFromRPC = true
	}
}

func WithDataVisualizationResponser(responser errcode.Responser) DataVisualizationOption {
	return func(o *dataVisualizationOptions) {
		o.responser = responser
	}
}

func WithDataVisualizationLogger(zapLog *zap.Logger) DataVisualizationOption {
	return func(o *dataVisualizationOptions) {
		o.zapLog = zapLog
	}
}

func WithDataVisualizationErrorToHTTPCode(e ...*errcode.Error) DataVisualizationOption {
	return func(o *dataVisualizationOptions) {
		o.httpErrors = e
	}
}

func WithDataVisualizationRPCStatusToHTTPCode(s ...*errcode.RPCStatus) DataVisualizationOption {
	return func(o *dataVisualizationOptions) {
		o.rpcStatus = s
	}
}

func WithDataVisualizationWrapCtx(wrapCtxFn func(c *gin.Context) context.Context) DataVisualizationOption {
	return func(o *dataVisualizationOptions) {
		o.wrapCtxFn = wrapCtxFn
	}
}

func RegisterDataVisualizationRouter(
	iRouter gin.IRouter,
	groupPathMiddlewares map[string][]gin.HandlerFunc,
	singlePathMiddlewares map[string][]gin.HandlerFunc,
	iLogic DataVisualizationLogicer,
	opts ...DataVisualizationOption) {

	o := &dataVisualizationOptions{}
	o.apply(opts...)

	if o.responser == nil {
		o.responser = errcode.NewResponser(o.isFromRPC, o.httpErrors, o.rpcStatus)
	}
	if o.zapLog == nil {
		o.zapLog, _ = zap.NewProduction()
	}

	r := &dataVisualizationRouter{
		iRouter:               iRouter,
		groupPathMiddlewares:  groupPathMiddlewares,
		singlePathMiddlewares: singlePathMiddlewares,
		iLogic:                iLogic,
		iResponse:             o.responser,
		zapLog:                o.zapLog,
		wrapCtxFn:             o.wrapCtxFn,
	}
	r.register()
}

type dataVisualizationRouter struct {
	iRouter               gin.IRouter
	groupPathMiddlewares  map[string][]gin.HandlerFunc
	singlePathMiddlewares map[string][]gin.HandlerFunc
	iLogic                DataVisualizationLogicer
	iResponse             errcode.Responser
	zapLog                *zap.Logger
	wrapCtxFn             func(c *gin.Context) context.Context
}

func (r *dataVisualizationRouter) register() {
	r.iRouter.Handle("POST", "/api/v1/dataVisualization/getDeviceData", r.withMiddleware("POST", "/api/v1/dataVisualization/getDeviceData", r.DeviceDataGet_0)...)
	r.iRouter.Handle("GET", "/api/v1/device/dataPush", r.withMiddleware("GET", "/api/v1/device/dataPush", r.DeviceDataPush_0)...)
	r.iRouter.Handle("POST", "/api/v1/dataVisualization/chart", r.withMiddleware("POST", "/api/v1/dataVisualization/chart", r.DeviceDataCurve_0)...)

}

func (r *dataVisualizationRouter) withMiddleware(method string, path string, fn gin.HandlerFunc) []gin.HandlerFunc {
	handlerFns := []gin.HandlerFunc{}

	// determine if a route group is hit or miss, left prefix rule
	for groupPath, fns := range r.groupPathMiddlewares {
		if groupPath == "" || groupPath == "/" {
			handlerFns = append(handlerFns, fns...)
			continue
		}
		size := len(groupPath)
		if len(path) < size {
			continue
		}
		if groupPath == path[:size] {
			handlerFns = append(handlerFns, fns...)
		}
	}

	// determine if a single route has been hit
	key := strings.ToUpper(method) + "->" + path
	if fns, ok := r.singlePathMiddlewares[key]; ok {
		handlerFns = append(handlerFns, fns...)
	}

	return append(handlerFns, fn)
}

var _ middleware.CtxKeyString

func (r *dataVisualizationRouter) DeviceDataGet_0(c *gin.Context) {
	req := &DeviceDataGetReq{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.DeviceDataGet(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *dataVisualizationRouter) DeviceDataPush_0(c *gin.Context) {
	req := &DeviceDataPushReq{}
	var err error

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.DeviceDataPush(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *dataVisualizationRouter) DeviceDataCurve_0(c *gin.Context) {
	req := &DeviceDataCurveReq{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.DeviceDataCurve(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}