# gateway/docs/apis.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "inner_core api docs",
    "version": "2.0"
  },
...
  "host": "127.0.0.1:8080",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/dataVisualization/chart": {
      "post": {
        "summary": "设备数据可视化",
        "description": "本接口为设备数据可视化接口",
        "operationId": "DataVisualization_DeviceDataCurve",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeviceDataCurveRes"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "description": "1. 用户ID",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "uint64"
          },
          {
            "name": "deviceCode",
            "description": "2. 设备编号",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "interval",
            "description": "3. 时间间隔\n\n - Null: 1. 空\n - Day: 2. 天\n - Week: 3. 星期\n - Month: 4 . 月",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "Null",
              "Day",
              "Week",
              "Month"
            ],
            "default": "Null"
          },
          {
            "name": "chartType",
            "description": "4. 图类型\n\n - ChartTypeNull: 1. 空\n - Pip: 1.饼图\n - Line: 2.折线图",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "ChartTypeNull",
                "Pip",
                "Line"
              ]
            },
            "collectionFormat": "multi"
          }
        ],
        "tags": [
          "DataVisualization"
        ]
      }
    },
    "/api/v1/dataVisualization/getDeviceData": {
      "post": {
        "summary": "查询设备数据",
        "description": "本接口为查询设备数据接口",
        "operationId": "DataVisualization_DeviceDataGet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeviceDataGetRes"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1DeviceDataGetReq"
            }
          }
        ],
        "tags": [
          "DataVisualization"
        ]
      }
    },
...
    "/api/v1/device/dataPush": {
      "get": {
        "summary": "设备数据推送",
        "description": "本接口为设备数据推送接口",
        "operationId": "DataVisualization_DeviceDataPush",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeviceDataPushRes"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "code",
            "description": "1. 设备编号",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "DataVisualization"
        ]
      }
    },
...
  },
  "definitions": {
    "commonDataDetail": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string",
          "title": "1. 键名称"
        },
        "value": {
          "type": "number",
          "format": "float",
          "title": "2. 键值"
        },
        "unit": {
          "type": "string",
          "title": "3. 单位"
        }
      }
    },
    "commonDeviceData": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "title": "1. 设备编号"
        },
        "data": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/commonDataDetail"
          },
          "title": "2. 详细数据"
        },
        "createTime": {
          "type": "string",
          "title": "3. 创建时间"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string",
          "description": "A URL/resource name that uniquely identifies the type of the serialized\nprotocol buffer message. This string must contain at least\none \"/\" character. The last segment of the URL's path must represent\nthe fully qualified name of the type (as in\n`path/google.protobuf.Duration`). The name should be in a canonical form\n(e.g., leading \".\" is not accepted).\n\nIn practice, teams usually precompile into the binary all types that they\nexpect it to use in the context of Any. However, for URLs which use the\nscheme `http`, `https`, or no scheme, one can optionally set up a type\nserver that maps type URLs to message definitions as follows:\n\n* If no scheme is provided, `https` is assumed.\n* An HTTP GET on the URL must yield a [google.protobuf.Type][]\n  value in binary format, or produce an error.\n* Applications are allowed to cache lookup results based on the\n  URL, or have them precompiled into a binary to avoid any\n  lookup. Therefore, binary compatibility needs to be preserved\n  on changes to types. (Use versioned type names to manage\n  breaking changes.)\n\nNote: this functionality is not currently available in the official\nprotobuf release, and it is not used for type URLs beginning with\ntype.googleapis.com.\n\nSchemes other than `http`, `https` (or the empty scheme) might be\nused with implementation specific semantics."
        }
      },
      "additionalProperties": {},
      "description": "`Any` contains an arbitrary serialized protocol buffer message along with a\nURL that describes the type of the serialized message.\n\nProtobuf library provides support to pack/unpack Any values in the form\nof utility functions or additional generated methods of the Any type.\n\nExample 1: Pack and unpack a message in C++.\n\n    Foo foo = ...;\n    Any any;\n    any.PackFrom(foo);\n    ...\n    if (any.UnpackTo(\u0026foo)) {\n      ...\n    }\n\nExample 2: Pack and unpack a message in Java.\n\n    Foo foo = ...;\n    Any any = Any.pack(foo);\n    ...\n    if (any.is(Foo.class)) {\n      foo = any.unpack(Foo.class);\n    }\n\n Example 3: Pack and unpack a message in Python.\n\n    foo = Foo(...)\n    any = Any()\n    any.Pack(foo)\n    ...\n    if any.Is(Foo.DESCRIPTOR):\n      any.Unpack(foo)\n      ...\n\n Example 4: Pack and unpack a message in Go\n\n     foo := \u0026pb.Foo{...}\n     any, err := anypb.New(foo)\n     if err != nil {\n       ...\n     }\n     ...\n     foo := \u0026pb.Foo{}\n     if err := any.UnmarshalTo(foo); err != nil {\n       ...\n     }\n\nThe pack methods provided by protobuf library will by default use\n'type.googleapis.com/full.type.name' as the type URL and the unpack\nmethods only use the fully qualified type name after the last '/'\nin the type URL, for example \"foo.bar.com/x/y.z\" will yield type\nname \"y.z\".\n\n\nJSON\n====\nThe JSON representation of an `Any` value uses the regular\nrepresentation of the deserialized, embedded message, with an\nadditional field `@type` which contains the type URL. Example:\n\n    package google.profile;\n    message Person {\n      string first_name = 1;\n      string last_name = 2;\n    }\n\n    {\n      \"@type\": \"type.googleapis.com/google.profile.Person\",\n      \"firstName\": \u003cstring\u003e,\n      \"lastName\": \u003cstring\u003e\n    }\n\nIf the embedded message type is well-known and has a custom JSON\nrepresentation, that representation will be embedded adding a field\n`value` which holds the custom JSON in addition to the `@type`\nfield. Example (for message [google.protobuf.Duration][]):\n\n    {\n      \"@type\": \"type.googleapis.com/google.protobuf.Duration\",\n      \"value\": \"1.212s\"\n    }"
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1DataLine": {
      "type": "object",
      "properties": {
        "key": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "1. 名称"
        },
        "value": {
          "type": "array",
          "items": {
            "type": "number",
            "format": "float"
          },
          "title": "2. 值"
        }
      }
    },
...
    "v1DeviceDataCurveRes": {
      "type": "object",
      "properties": {
        "month": {
          "$ref": "#/definitions/v1DataLine",
          "title": "1. 近一个月数据量"
        },
        "pip": {
          "$ref": "#/definitions/v1DataLine",
          "title": "2. 饼图"
        },
        "line": {
          "$ref": "#/definitions/v1DataLine",
          "title": "3. 折线图"
        },
        "elect": {
          "$ref": "#/definitions/v1DataLine",
          "title": "4. 电流"
        },
        "volt": {
          "$ref": "#/definitions/v1DataLine",
          "title": "5. 电压"
        }
      }
    },
    "v1DeviceDataGetReq": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "1. 设备名称"
        },
        "code": {
          "type": "string",
          "title": "2. 设备编号"
        },
        "pageSize": {
          "type": "integer",
          "format": "uint64",
          "title": "3. 页大小"
        },
        "startTime": {
          "type": "string",
          "title": "4. 开始时间"
        },
        "endTime": {
          "type": "string",
          "title": "5. 结束时间"
        },
        "currPage": {
          "type": "integer",
          "format": "uint64",
          "title": "6. 页码"
        },
        "userId": {
          "type": "integer",
          "format": "uint64",
          "title": "7. 用户ID"
        }
      }
    },
    "v1DeviceDataGetRes": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "1. 设备名称"
        },
        "code": {
          "type": "string",
          "title": "2. 设备编号"
        },
        "data": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/commonDeviceData"
          },
          "title": "3. 设备数据"
        },
        "count": {
          "type": "integer",
          "format": "uint64",
          "title": "4. 数据量"
        }
      }
    },
    "v1DeviceDataPushRes": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "1. 设备名称"
        },
        "code": {
          "type": "string",
          "title": "2. 设备编号"
        },
        "data": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/commonDataDetail"
          },
          "title": "3. 设备数据"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "description": "Input a \"Bearer your-jwt-token\" to Value",
      "name": "Authorization",
      "in": "header"
    }
  }
}